Feat (Logic Analyzer): Auto Recapture and Capture Time slider
Capture time slider is changed to allow for larger values and more
fine-tuned control. Auto Recapture checkbox added to have the logic
analyzer continously recapture until an event is detected.

The Graph component is recieving the LA data and if a capture has no
event within it and the auto recapture is on, the LA is toggled
back on.

Standarized UI to use μs which is converted to ms for the backend communication.

Fix (Logic Analyzer): Define inputCheck

The CustomDialog dispatch needs an inputCheck function to be defined.
The Logic Analyzer capture time slider will always be an
non-negative float so inputCheck will check for that.